pgsql: fix type of mac_protocol and oob_protocol.
The oob_protocol does not fit into a smallint (IPv6 case for example).
This patch switches the length of the related field to an integer.

+++ doc/pgsql-ulogd2-flat.sql
@@ -52,7 +52,7 @@ CREATE TABLE ulog2 (
   -- timestamp timestamp NOT NULL default 'now',
   mac_saddr_str macaddr default NULL,
   mac_daddr_str macaddr default NULL,
-  oob_protocol smallint default NULL,
+  oob_protocol integer default NULL,
   raw_type integer default NULL,
   mac_str varchar(256) default NULL,
   tcp_sport integer default NULL,
